SEQNG-244 Improved display of the sequence config #time 20m
